Compute partial pathsums
Instead of recomputing the entire minpath pathsum map, only recompute values
that were changed. Start with recomputing the two possibly changed values in
the 2nd row, and expand outwards in a triangle when moving down. If the seam is
approximately vertical, this requires about half as many computations as
before.